[REFACTOR] declare ‘Do’ scope as helper
after trying to fit the ‘Do’ on service level, realized that it’s not a service really (since servcie is supposed to hold at least some internal state, hence should be an instantiatable reference type, i.e. class), it’s a helper/alyas, and can be accessed from any layer (this particular is supposed to be used from Controller level).