[REFACTOR] naming improvements around ‘App’ vs. ‘AppDelegate’
renamed the ‘App’ type into ‘App_Ctr’ to emphasise that the ‘@UIApplicationMain’ is actually nothin more than a special kind of controller that exists in any app and controls the app in general. It also has to conform to ‘UIAppDelegate’, which requires it to have the ‘window’ property, which, in turn, plays same role for the app-wide controller as the ‘view’ property for regular UIViewController-based controllers. Also re-introduced ‘App’ type as scope (special case of Component type that does not need to hold any state) for consistancy (because every Controller has to belong to a particular scope/Component).
NOTE: we kept the ‘AppDelegate’ file at the root level of ‘Sources’ folder/group to make it easier for anyone to find the entry point in the app, since that’s the name of the file where Xcode projects historically had app initialization code.